+++ __tests__/JS_tests.js
@@ -14,19 +14,14 @@
  *   limitations under the License.
  */
 
-const {
-  compile,
-  makeAst,
-  renderContext,
-  renderContextAsync,
-} = require("../src/AcutisJs.bs");
+const { Compile, Environment } = require("../src/AcutisJs.bs");
 
 describe("The JS interface works as expected", () => {
   test("Bad input is reported correctly", () => {
-    function X(render, props, children) {
-      return render(makeAst(1), props, children);
+    function X(env, props, children) {
+      return env.render(Compile.makeAst(1, "X"), props, children);
     }
-    expect(X(renderContext({}), { name: "Carlo" })).toEqual({
+    expect(X(Environment.make({}), { name: "Carlo" })).toEqual({
       NAME: "errors",
       VAL: [
         {
@@ -37,18 +32,21 @@ describe("The JS interface works as expected", () => {
             RE_EXN_ID: "Caml_js_exceptions.Error/2",
             _1: new TypeError("source.str.charAt is not a function"),
           },
+          path: ["X"],
         },
       ],
     });
-    function Y(render, props, children) {
-      return render("{{ name }}", props, children);
+    function Y(env, props, children) {
+      return env.render("{{ name }}", props, children);
     }
-    expect(Y(renderContext({}), { name: "Carlo" }, {})).toEqual({
+    expect(Y(Environment.make({}), { name: "Carlo" }, {})).toEqual({
       NAME: "errors",
       VAL: [
         {
           kind: "Render",
-          message: "An AST was not valid. Did you forget to compile it?",
+          message:
+            "A template AST was not valid. Did you forget to compile one?",
+          path: [],
         },
       ],
     });
@@ -56,81 +54,160 @@ describe("The JS interface works as expected", () => {
 });
 
 describe("Async templates", () => {
-  test("Async context", async () => {
-    const X = compile("{{ name }}", "X");
-    const renderContext = renderContextAsync({});
-    const x = await X(renderContext, { name: "Carlo" }, {});
+  test("Async env", async () => {
+    const X = Compile.make("{{ name }}", "X");
+    const env = Environment.Async.make({});
+    const x = await X(env, { name: "Carlo" }, {});
     expect(x).toEqual({ NAME: "data", VAL: "Carlo" });
   });
 
   test("Async template components", async () => {
-    const yAst = makeAst("{{ name }}");
-    async function Y(render, props, children) {
-      return render(yAst, props, children);
+    const yAst = Compile.makeAst("{{ name }}");
+    async function Y(env, props, children) {
+      return env.render(yAst, props, children);
     }
-    const X = compile("{% Y name /%}", "X");
-    const renderContext = renderContextAsync({ Y: Y });
-    const x = await X(renderContext, { name: "Carlo" }, {});
+    const X = Compile.make("{% Y name /%}", "X");
+    const env = Environment.Async.make({ Y: Y });
+    const x = await X(env, { name: "Carlo" }, {});
     expect(x).toEqual({ NAME: "data", VAL: "Carlo" });
   });
 
   test("Async template component errors", async () => {
-    async function Y(render, props, children) {
-      return render(123, props, children);
+    async function Y(env, props, children) {
+      return env.render(123, props, children);
     }
-    const X = compile("{% Y name /%}", "X");
-    const renderContext = renderContextAsync({ Y: Y });
-    expect(await X(renderContext, { name: "Carlo" }, {})).toEqual({
+    const X = Compile.make("{% Y name /%}", "X");
+    const env = Environment.Async.make({ Y: Y });
+    expect(await X(env, { name: "Carlo" }, {})).toEqual({
       NAME: "errors",
       VAL: [
         {
           kind: "Render",
-          message: "An AST was not valid. Did you forget to compile it?",
+          message:
+            "A template AST was not valid. Did you forget to compile one?",
+          path: ["X"],
         },
       ],
     });
-    async function A(render, props, children) {
-      return render(makeAst("{{ name }}", "A"), props, children);
+    async function A(env, props, children) {
+      return env.render(Compile.makeAst("{{ name }}", "A"), props, children);
     }
-    const B = compile("{% A /%}", "B");
-    const renderContext2 = renderContextAsync({ A: A });
-    expect(await B(renderContext2, {}, {})).toEqual({
+    const B = Compile.make("{% A /%}", "B");
+    const env2 = Environment.Async.make({ A: A });
+    expect(await B(env2, {}, {})).toEqual({
       NAME: "errors",
       VAL: [
         {
           message: '"name" is type null. I can only echo strings and numbers.',
-          template: "A",
+          path: ["A", "B"],
           location: { character: 4 },
           kind: "Render",
         },
       ],
     });
-    const C = compile("{{ a", "C");
-    const renderContext3 = renderContextAsync({});
-    expect(await C(renderContext3, {}, {})).toEqual({
+    const C = Compile.make("{{ a", "C");
+    const env3 = Environment.Async.make({});
+    expect(await C(env3, {}, {})).toEqual({
       NAME: "errors",
       VAL: [
         {
           message: 'Unexpected character: "". Expected: "}}".',
-          template: "C",
           location: { character: 7 },
           kind: "Syntax",
+          path: ["C"],
         },
       ],
     });
-    async function D(render, props, children) {
+    async function D(_env, _props, _children) {
       throw new Error("fail.");
     }
-    const E = compile("{% D /%}", "E");
-    const renderContext4 = renderContextAsync({ D: D });
-    expect(await E(renderContext4, {}, {})).toEqual({
+    const E = Compile.make("{% D /%}", "E");
+    const env4 = Environment.Async.make({ D: D });
+    expect(await E(env4, {}, {})).toEqual({
       NAME: "errors",
       VAL: [
         {
-          message: `An exception was thrown while rendering this template. This is probably due to malformed input.`,
+          message: `An exception was thrown while rendering a template component.`,
           kind: "Render",
-          template: "D",
+          path: ["E"],
           exn: new Error("fail."),
+        },
+      ],
+    });
+  });
+});
+
+describe("Async helper functions", () => {
+  test("env.return", async () => {
+    const X = (env, _props, _children) => {
+      return env.return("a");
+    };
+    const Template = Compile.make("{% X / %}");
+    const env = Environment.Async.make({ X: X });
+    expect(await Template(env, {}, {})).toEqual({ NAME: "data", VAL: "a" });
+  });
+
+  test("env.error", async () => {
+    const X = (env, _props, _children) => {
+      return env.error("e");
+    };
+    const Template = Compile.make("{% X / %}");
+    const env = Environment.Async.make({ X: X });
+    expect(await Template(env, {}, {})).toEqual({
+      NAME: "errors",
+      VAL: [
+        {
+          message: "e",
+          kind: "Render",
+          path: [null],
+        },
+      ],
+    });
+  });
+
+  test("env.mapChild", async () => {
+    const X = (env, _props, { Children }) => {
+      return env.mapChild(Children, (x) => x.toUpperCase());
+    };
+    const env = Environment.Async.make({ X: X });
+    const Template = Compile.make("{% X ~%} a {%~ /X %}");
+    expect(await Template(env, {}, {})).toEqual({
+      NAME: "data",
+      VAL: "A",
+    });
+    const BadTemplate = Compile.make("{% X %} {{ e }} {% /X %}", "BadTemplate");
+    expect(await BadTemplate(env, {}, {})).toEqual({
+      NAME: "errors",
+      VAL: [
+        {
+          message: '"e" is type null. I can only echo strings and numbers.',
+          location: { character: 12 },
+          kind: "Render",
+          path: ["section: X#Children", "BadTemplate"],
+        },
+      ],
+    });
+  });
+
+  test("env.flatMapChild", async () => {
+    const X = (env, _props, { Children }) => {
+      return env.flatMapChild(Children, (x) => env.return(x.toUpperCase()));
+    };
+    const env = Environment.Async.make({ X: X });
+    const Template = Compile.make("{% X %} a {% /X %}");
+    expect(await Template(env, {}, {})).toEqual({
+      NAME: "data",
+      VAL: " A ",
+    });
+    const BadTemplate = Compile.make("{% X %} {{ e }} {% /X %}", "BadTemplate");
+    expect(await BadTemplate(env, {}, {})).toEqual({
+      NAME: "errors",
+      VAL: [
+        {
+          message: '"e" is type null. I can only echo strings and numbers.',
+          location: { character: 12 },
+          kind: "Render",
+          path: ["section: X#Children", "BadTemplate"],
         },
       ],
     });